# api/api-spec.yaml
openapi: 3.0.3
info:
  title: TripleThink API
  description: |
    Event-sourced narrative construction system for multi-book fiction series.

    TripleThink enables authors to manage complex storylines with:
    - Epistemic tracking (who knows what, when)
    - Fiction management (false narratives with strict audience constraints)
    - Time-travel queries (entity state at any timestamp)
    - Token-efficient AI integration

    ## Key Concepts

    - **Entities**: Characters, events, objects, locations, systems
    - **Metadata**: Rich author notes and AI guidance (loaded on-demand)
    - **Knowledge States**: What characters believe at specific times
    - **Fictions**: False narratives with target audiences
    - **Epistemic Queries**: The power feature for tracking beliefs
  version: 1.0.0
  contact:
    name: TripleThink Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/entities:
    get:
      tags: [Entities]
      summary: List entities
      parameters:
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/EntityType'
        - name: date_range
          in: query
          description: 'Format: start,end (ISO8601)'
          schema:
            type: string
            example: '2033-01-01T00:00:00Z,2033-12-31T23:59:59Z'
        - name: participant
          in: query
          description: Filter events by participant character ID
          schema:
            type: string
        - name: search
          in: query
          description: Search in name and summary
          schema:
            type: string
        - name: include_metadata
          in: query
          schema:
            $ref: '#/components/schemas/MetadataMode'
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of entities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entity'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Entities]
      summary: Create entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, data]
              properties:
                type:
                  $ref: '#/components/schemas/EntityType'
                data:
                  type: object
                  required: [id, name]
                  properties:
                    id:
                      type: string
                      example: char-new-character
                    name:
                      type: string
                      example: New Character
                metadata:
                  $ref: '#/components/schemas/Metadata'
      responses:
        '201':
          description: Entity created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Entity'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /api/export/project:
    get:
      tags: [Export/Import]
      summary: Export complete project
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [json]
            default: json
        - name: include_metadata
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Project export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectExport'
  /api/export/entities:
    post:
      tags: [Export/Import]
      summary: Export specific entities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_ids]
              properties:
                entity_ids:
                  type: array
                  items:
                    type: string
                include_metadata:
                  type: boolean
                  default: true
                include_related:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Partial export
  /api/import/project:
    post:
      tags: [Export/Import]
      summary: Import project data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  $ref: '#/components/schemas/ProjectExport'
                mode:
                  type: string
                  enum: [merge, skip, replace]
                  default: merge
                dry_run:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Import result

  # ==========================================
  # SEARCH
  # ==========================================
...
components:
  schemas:
    EntityType:
      type: string
      enum: [event, character, object, location, system, fiction]
    MetadataMode:
      type: string
      enum: [auto, always, never]
      default: auto
    Entity:
      type: object
      properties:
        id:
          type: string
        entity_type:
          $ref: '#/components/schemas/EntityType'
        name:
          type: string
        timestamp:
          type: string
          format: date-time
        summary:
          type: string
        data:
          type: object
        meta_id:
          type: string
        read_metadata_mandatory:
          type: boolean
        metadata:
          $ref: '#/components/schemas/Metadata'
    Metadata:
      type: object
      properties:
        id:
          type: string
        entity_id:
          type: string
        entity_type:
          type: string
        author_notes:
          type: object
          properties:
            creative_intent:
              type: string
            themes:
              type: array
              items:
                type: string
            constraints:
              type: array
              items:
                type: string
        ai_guidance:
          type: object
          properties:
            tone:
              type: string
            scene_rendering:
              type: string
            pov_recommendations:
              type: array
              items:
                type: string
        dev_status:
          type: object
          properties:
            completeness:
              type: string
            todo:
              type: array
              items:
                type: string
            warnings:
              type: array
              items:
                type: string
        version_info:
          type: object
          required: [created, modified]
          properties:
            created:
              type: string
              format: date-time
            modified:
              type: string
              format: date-time
            changelog:
              type: array
              items:
                type: object
        prose_guidance:
          type: object
          properties:
            voice:
              type: string
            tone:
              type: string
            pacing_notes:
              type: string
        consistency_rules:
          type: array
          items:
            type: string
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        count:
          type: integer
        total:
          type: integer
    ProjectExport:
      type: object
      properties:
        export_info:
          type: object
          properties:
            format:
              type: string
            exported_at:
              type: string
              format: date-time
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/Metadata'
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            suggestion:
              type: string
            timestamp:
              type: string
              format: date-time
  responses:
    ValidationError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Missing required fields
              details:
                missing_fields: [id, name]
              suggestion: Check request body for required fields
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []